push

+++ controller/update_interract.php
@@ -12,8 +12,10 @@
       $login = $db->query("SELECT login FROM user WHERE id=" . $_SESSION['id'])->fetchAll()[0]['login'];
       $return = "<span class='pseudo'>" . $login . "</span>";
       $return .= "<span class='value'>" . $ret . "</span>";
-      $return .= "<a href='#'>modifier</a>";
-      $return .= "<a href='#'>supprimer</a>";
+      $return .= "<span class='hidded'>";
+        $return .= "<a href='#'>modifier</a>";
+        $return .= "<a href='#'>supprimer</a>";
+      $return .= "</span>";
       return ($return);
     }
     else
@@ -22,39 +24,44 @@
 
   if (isset($_POST['id']) && isset($_POST['state']) && isset($_POST['type']))
   {
-    $id = htmlspecialchars($_POST['id']);
-    $state = htmlspecialchars($_POST['state']);
-    $type = htmlspecialchars($_POST['type']);
-    $id_interract = htmlspecialchars($_POST['id_interract']);
-
-    if ($type == "like")
+    if (isset($_SESSION['id']))
     {
-      if ($state)
-        echo $db->exec("DELETE FROM interract WHERE id=" . $id_interract);
-      else
+      $id = htmlspecialchars($_POST['id']);
+      $state = htmlspecialchars($_POST['state']);
+      $type = htmlspecialchars($_POST['type']);
+      $id_interract = htmlspecialchars($_POST['id_interract']);
+
+      if ($type == "like")
       {
+        if ($state)
+          echo $db->exec("DELETE FROM interract WHERE id=" . $id_interract);
+        else
+        {
+          $params = array(
+            ':id_picture' => $id,
+            ':id_user' => $_SESSION['id'],
+            ':type' => $type,
+            ':value' => '0'
+          );
+          echo insert_interract($params, $db);
+        }
+      }
+      else if ($type == "comment")
+      {
+        $val = htmlspecialchars($_POST['value']);
         $params = array(
           ':id_picture' => $id,
           ':id_user' => $_SESSION['id'],
           ':type' => $type,
-          ':value' => '0'
+          ':value' => $val
         );
-        echo insert_interract($params, $db);
+        echo insert_interract($params, $db, $val);
       }
     }
-    else if ($type == "comment")
-    {
-      $val = htmlspecialchars($_POST['value']);
-      $params = array(
-        ':id_picture' => $id,
-        ':id_user' => $_SESSION['id'],
-        ':type' => $type,
-        ':value' => $val
-      );
-      echo insert_interract($params, $db, $val);
-    }
+    else
+      echo "-2";
   }
   else
-    echo "bad request send";
+    echo "-1";
 
 ?>
